Refactor: move _run_analysis to services/analysis_runner.py
- Extract _run_analysis, _extract_input_data, _to_linear_data,
  route_level1_decision, EXPECTED_STRUCTURE from routes/analyze.py
- Wrap as Celery task (run_analysis_task) with bind=True
- analyze.py now imports _run_analysis from analysis_runner;
  submit() call signature unchanged
- Update test_analyze_routing.py import path accordingly
