+++ backend/requirements.txt
@@ -17,3 +17,6 @@ scipy>=1.12.0
 fastembed>=0.4.0,<1
 google-genai>=1.73.0
 httpx>=0.27.0
+celery>=5.6.3
+redis>=7.4
+limits[redis]>=5.8
